backend/requirements.txt
# FastAPI and server
fastapi==0.115.9
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Database
pymilvus==2.6.6

# Transformers and embeddings
transformers==4.46.0
peft==0.13.2
FlagEmbedding==1.2.11
open_clip_torch==2.26.1
protobuf>=5.27.2
sentencepiece
boto3

# Video processing
yt-dlp==2024.12.23
opencv-python-headless>=4.10.0.84

# Utilities
numpy>=2.1.0,<2.3.0
pandas==2.2.3
pillow==10.4.0
tqdm==4.66.5

# Configuration
pydantic==2.9.0
pydantic-settings==2.5.2
python-dotenv==1.0.1

# HTTP client
httpx==0.27.2
aiofiles==24.1.0